+++ packages/app-utils/src/files/index.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-export function mkdirp(dir) {
+export function mkdirp(dir: string) {
 	try {
 		fs.mkdirSync(dir, { recursive: true });
 	} catch (e) {
@@ -10,25 +10,21 @@ export function mkdirp(dir) {
 	}
 }
 
-export function copy (
-	from,
-	to,
-	filter: (file?: string) => boolean = () => true
-):string[] {
-	if (!filter(path.basename(from))) return;
+export function copy(from: string, to: string, filter: (file?: string) => boolean = () => true): string[] {
+	if (!filter(path.basename(from))) return [];
 
-  const files = []
+	const files = [];
 	const stats = fs.statSync(from);
 
 	if (stats.isDirectory()) {
 		fs.readdirSync(from).forEach(file => {
-      files.push(...copy(path.join(from, file), path.join(to, file)));
-    });
+			files.push(...copy(path.join(from, file), path.join(to, file)));
+		});
 	} else {
 		mkdirp(path.dirname(to));
-    fs.copyFileSync(from, to);
-    files.push(to);
-  }
-  
-  return files;
+		fs.copyFileSync(from, to);
+		files.push(to);
+	}
+
+	return files;
 }