+++ packages/app-utils/src/index.ts
@@ -1,3 +1,0 @@
-export { render } from './render';
-export { prerender } from './prerender';
-export * from './files';